// frontend/src/interfaces/pet.ts
export interface Pet {
  id: string
  name: string
  gender: string
  size: string
  user_id: string
  pet_photos: string
  history: string
  castrated: boolean
  vaccinated: boolean
  city: string
  uf: string
  pet_location: string
  pet_owner_email: string
  specie: string
  createdAt: string
  updatedAt: string
}